move det code to adapter
